Update online.rst
IRC details removed. It's the 21st century.
Substantive changes to layout and content
URLs moved to titles for ease of reference, descriptions redone to fit the purpose.
Syntax, grammar and tone.
Ordering changed to bring more relevant information to the top (homepage and HMS).

Would be good to get a newsletter.nottinghack.org.uk URL for the newsletter signup
Is it possible to cross-reference team names to the relevant team details later in the guide via hyperlink?
Need to fix hyperlinks for links that don't start www.